Use reflog for old_ref.
ORIG_HEAD is only correct when doing dangerous operations like merging/cherry-picking/reverting/bisecting/rebasing.
But I am guessing that using git_rails with no arguments will most often be after a successful operation that registers in the reflog.
This won't work if you are in a merge/cherry-pick/revert conflict or are bisecting/rebasing interactively.