Update the docs.
I wish I knew of a good way of checking whether we are in a situation
where `ORIG_HEAD` is better to use than `HEAD@{1}`

From looking at the prompt implementation `__git_ps1` it doesn't look
like there is an easy way to detect when to use `ORIG_HEAD` instead of
`reflog`.
That implementation looks for the existence of several files
`.git/{MERGE_HEAD,CHERRY_PICK_HEAD,REVERT_HEAD,BISECT_LOG,rebase-merge/interactive}`
which I could do, but I think it is a bit fragile, and I wish I knew of
a better way.